+++ CHANGELOG.md
@@ -6,7 +6,10 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 
 ## [unreleased]
 
-\-
+### Added
+- Increment tripcode in New collecting event task [#2441]
+
+[#2441]: https://github.com/SpeciesFileGroup/taxonworks/issues/2441
 
 ## [0.24.1] - 2022-04-04